// C0920G1-CarInsurance-FE/src/app/models/contract/accident-dto.ts
export class AccidentDTO {

   customerId: string;
   customerName: string;
   customerAddress: string;
   numberPlate: string;
   carIdNumber: string;
   manufacturer: string;
   accidentReason: string;
   accidentImg: string;
   accidentMoney: string;
   notApprovalReason: string;
   productPrice: string;
   customerMail: string;
   contractId:string;
}